+++ src/grammar/grammar.ts
@@ -823,6 +823,34 @@ semantics.addOperation<ASTNode>("resolve_statement", {
             ref: createRef(this),
         });
     },
+    StatementTry_simple(_arg0, _arg1, arg2, _arg3) {
+        return createNode({
+            kind: "statement_try",
+            statements: arg2.children.map((v) => v.resolve_statement()),
+            ref: createRef(this),
+        });
+    },
+    StatementTry_withCatch(
+        _arg0,
+        _arg1,
+        arg2,
+        _arg3,
+        _arg4,
+        _arg5,
+        arg6,
+        _arg7,
+        _arg8,
+        arg9,
+        _arg10,
+    ) {
+        return createNode({
+            kind: "statement_try_catch",
+            statements: arg2.children.map((v) => v.resolve_statement()),
+            catchName: arg6.sourceString,
+            catchStatements: arg9.children.map((v) => v.resolve_statement()),
+            ref: createRef(this),
+        });
+    },
 });
 
 // LValue